Load graphs asynchronously to lighten page size
